replace x,y,z to real x86 regs

+++ stack.h
@@ -1,5 +1,5 @@
 #ifndef _STACK_H_
 #define _STACK_H_
-extern void push_address(unsigned long);
-unsigned long pop_address(void);
+extern void push_address(unsigned long long);
+unsigned long long pop_address(void);
 #endif
